Split RC adjustments into separate module
